Bug #1257: compile warnings due to pid_t not being an int

+++ helpers/ntlm_auth/winbind/wbntlm.h
@@ -48,7 +48,7 @@ extern char debug_enabled;
 #endif
 
 #if defined(__GNUC__) || defined(__ICC)    /* this is really a gcc-ism */
-#define warn(X...)  fprintf(stderr,"%s[%d](%s:%d): ", myname, mypid, \
+#define warn(X...)  fprintf(stderr,"%s[%ld](%s:%d): ", myname, (long)mypid, \
                     ((__foo=strrchr(__FILE__,'/'))==NULL?__FILE__:__foo+1),\
                     __LINE__);\
                     fprintf(stderr,X)
